[Draft] Splitted object Text and Label from Draft.py
And based them on Draft Annotation object
